Add project owner filter to worktable
- Also refactor worktable summary row
