+++ backend/src/router/workTable.ts
@@ -1,5 +1,4 @@
 import { TRPCError } from '@trpc/server';
-import { hasWritePermission, ownsProject } from '@shared/schema/userPermissions';
 
 import { textToSearchTerms } from '@backend/components/project/search';
 import { getPool, sql } from '@backend/db';
@@ -8,6 +7,7 @@ import { getProjectObjects, upsertProjectObject } from '@backend/router/projectO
 
 import { UpsertProjectObject } from '@shared/schema/projectObject';
 import { User } from '@shared/schema/user';
+import { hasWritePermission, ownsProject } from '@shared/schema/userPermissions';
 import {
   WorkTableSearch,
   WorkTableUpdate,
@@ -33,6 +33,7 @@ async function workTableSearch(input: WorkTableSearch) {
     objectUsage = [],
     lifecycleState = [],
     objectStage = [],
+    projectOwner = null,
   } = input;
 
   const query = sql.type(workTableRowSchema)`
@@ -45,6 +46,7 @@ async function workTableSearch(input: WorkTableSearch) {
     INNER JOIN app.project_investment ON project_investment.id = project.id
 
     WHERE project_object.deleted = false
+      ${projectOwner ? sql.fragment`AND owner = ${projectOwner}` : sql.fragment``}
       -- search date range intersection
       AND daterange(${startDate}, ${endDate}, '[]') && daterange(project_object.start_date, project_object.end_date, '[]')
       AND (${objectNameSearch}::text IS NULL OR to_tsquery('simple', ${objectNameSearch}) @@ to_tsvector('simple', project_object.object_name) OR project_object.object_name LIKE '%' || ${objectNameSubstringSearch} || '%')
